Moves all actions attached to the HTTP method `GET` to the namespace `Api\Get`.
