+++ src/Actions/Api/Get/GetCurrentTrackAction.php
@@ -1,5 +1,5 @@
 <?php declare( strict_types = 1 );
-namespace CodeKandis\TradioApi\Actions\Api\Read;
+namespace CodeKandis\TradioApi\Actions\Api\Get;
 
 use CodeKandis\Tiphy\Actions\AbstractAction;
 use CodeKandis\Tiphy\Exceptions\ErrorInformation;
@@ -9,17 +9,19 @@ use CodeKandis\Tiphy\Persistence\MariaDb\Connector;
 use CodeKandis\Tiphy\Persistence\MariaDb\ConnectorInterface;
 use CodeKandis\Tiphy\Persistence\PersistenceException;
 use CodeKandis\TradioApi\Configurations\ConfigurationRegistry;
+use CodeKandis\TradioApi\Entities\CurrentTrackEntity;
 use CodeKandis\TradioApi\Entities\FavoriteEntity;
 use CodeKandis\TradioApi\Entities\StationEntity;
-use CodeKandis\TradioApi\Entities\UriExtenders\FavoriteUriExtender;
+use CodeKandis\TradioApi\Entities\UriExtenders\CurrentTrackUriExtender;
 use CodeKandis\TradioApi\Errors\StationsErrorCodes;
 use CodeKandis\TradioApi\Errors\StationsErrorMessages;
+use CodeKandis\TradioApi\Http\Readers\CurrentTrackReader;
 use CodeKandis\TradioApi\Http\UriBuilders\ApiUriBuilder;
 use CodeKandis\TradioApi\Persistence\MariaDb\Repositories\FavoritesRepository;
 use CodeKandis\TradioApi\Persistence\MariaDb\Repositories\StationsRepository;
 use ReflectionException;
 
-class GetStationFavoritesAction extends AbstractAction
+class GetCurrentTrackAction extends AbstractAction
 {
 	/** @var ConnectorInterface */
 	private $databaseConnector;
@@ -70,12 +72,16 @@ class GetStationFavoritesAction extends AbstractAction
 			return;
 		}
 
-		$favorites = $this->readStationsFavorites( $station );
-		$this->extendUris( $favorites );
+		$currentTrack            = $this->readCurrentTrack( $station );
+		$requestedFavorite       = new FavoriteEntity();
+		$requestedFavorite->name = $currentTrack->name;
+		$favorite                = $this->readFavoriteByTrackName( $requestedFavorite );
+		$this->extendUris( $currentTrack, $station, $favorite );
 
 		$responderData = [
-			'favorites' => $favorites,
+			'currentTrack' => $currentTrack
 		];
+
 		( new JsonResponder( StatusCodes::OK, $responderData ) )
 			->respond();
 	}
@@ -88,17 +94,11 @@ class GetStationFavoritesAction extends AbstractAction
 		return $this->arguments;
 	}
 
-	/**
-	 * @param FavoriteEntity[] $favorites
-	 */
-	private function extendUris( array $favorites ): void
+	private function extendUris( CurrentTrackEntity $currentTrack, StationEntity $station, ?FavoriteEntity $favorite ): void
 	{
 		$uriBuilder = $this->getUriBuilder();
-		foreach ( $favorites as $favorite )
-		{
-			( new FavoriteUriExtender( $uriBuilder, $favorite ) )
-				->extend();
-		}
+		( new CurrentTrackUriExtender( $uriBuilder, $currentTrack, $station, $favorite ) )
+			->extend();
 	}
 
 	/**
@@ -112,15 +112,24 @@ class GetStationFavoritesAction extends AbstractAction
 			->readStationById( $requestedStation );
 	}
 
+	private function readCurrentTrack( StationEntity $station ): CurrentTrackEntity
+	{
+		$currentTrack            = new CurrentTrackEntity();
+		$currentTrack->name      = ( new CurrentTrackReader() )
+			->read( $station->tracklistUri, $station->currentTrackXPath );
+		$currentTrack->stationId = $station->id;
+
+		return $currentTrack;
+	}
+
 	/**
-	 * @return FavoriteEntity[]
 	 * @throws PersistenceException
 	 */
-	private function readStationsFavorites( StationEntity $station ): array
+	private function readFavoriteByTrackName( FavoriteEntity $requestedFavorite ): ?FavoriteEntity
 	{
 		$databaseConnector = $this->getDatabaseConnector();
 
 		return ( new FavoritesRepository( $databaseConnector ) )
-			->readFavoritesByStationId( $station );
+			->readFavoriteByTrackName( $requestedFavorite );
 	}
 }
