+++ src/Actions/Api/Get/GetFavoriteStationsAction.php
@@ -1,5 +1,5 @@
 <?php declare( strict_types = 1 );
-namespace CodeKandis\TradioApi\Actions\Api\Read;
+namespace CodeKandis\TradioApi\Actions\Api\Get;
 
 use CodeKandis\Tiphy\Actions\AbstractAction;
 use CodeKandis\Tiphy\Exceptions\ErrorInformation;
@@ -9,19 +9,17 @@ use CodeKandis\Tiphy\Persistence\MariaDb\Connector;
 use CodeKandis\Tiphy\Persistence\MariaDb\ConnectorInterface;
 use CodeKandis\Tiphy\Persistence\PersistenceException;
 use CodeKandis\TradioApi\Configurations\ConfigurationRegistry;
-use CodeKandis\TradioApi\Entities\CurrentTrackEntity;
 use CodeKandis\TradioApi\Entities\FavoriteEntity;
 use CodeKandis\TradioApi\Entities\StationEntity;
-use CodeKandis\TradioApi\Entities\UriExtenders\CurrentTrackUriExtender;
-use CodeKandis\TradioApi\Errors\StationsErrorCodes;
-use CodeKandis\TradioApi\Errors\StationsErrorMessages;
-use CodeKandis\TradioApi\Http\Readers\CurrentTrackReader;
+use CodeKandis\TradioApi\Entities\UriExtenders\StationUriExtender;
+use CodeKandis\TradioApi\Errors\FavoritesErrorCodes;
+use CodeKandis\TradioApi\Errors\FavoritesErrorMessages;
 use CodeKandis\TradioApi\Http\UriBuilders\ApiUriBuilder;
 use CodeKandis\TradioApi\Persistence\MariaDb\Repositories\FavoritesRepository;
 use CodeKandis\TradioApi\Persistence\MariaDb\Repositories\StationsRepository;
 use ReflectionException;
 
-class GetCurrentTrackAction extends AbstractAction
+class GetFavoriteStationsAction extends AbstractAction
 {
 	/** @var ConnectorInterface */
 	private $databaseConnector;
@@ -59,29 +57,25 @@ class GetCurrentTrackAction extends AbstractAction
 	{
 		$inputData = $this->getInputData();
 
-		$requestedStation     = new StationEntity();
-		$requestedStation->id = $inputData[ 'id' ];
-		$station              = $this->readStation( $requestedStation );
+		$requestedFavorite     = new FavoriteEntity();
+		$requestedFavorite->id = $inputData[ 'id' ];
+		$favorite              = $this->readFavorite( $requestedFavorite );
 
-		if ( null === $station )
+		if ( null === $favorite )
 		{
-			$errorInformation = new ErrorInformation( StationsErrorCodes::STATION_UNKNOWN, StationsErrorMessages::STATION_UNKNOWN, $inputData );
+			$errorInformation = new ErrorInformation( FavoritesErrorCodes::FAVORITE_UNKNOWN, FavoritesErrorMessages::FAVORITE_UNKNOWN, $inputData );
 			( new JsonResponder( StatusCodes::NOT_FOUND, null, $errorInformation ) )
 				->respond();
 
 			return;
 		}
 
-		$currentTrack            = $this->readCurrentTrack( $station );
-		$requestedFavorite       = new FavoriteEntity();
-		$requestedFavorite->name = $currentTrack->name;
-		$favorite                = $this->readFavoriteByTrackName( $requestedFavorite );
-		$this->extendUris( $currentTrack, $station, $favorite );
+		$stations = $this->readFavoriteStations( $favorite );
+		$this->extendUris( $stations );
 
 		$responderData = [
-			'currentTrack' => $currentTrack
+			'stations' => $stations,
 		];
-
 		( new JsonResponder( StatusCodes::OK, $responderData ) )
 			->respond();
 	}
@@ -94,42 +88,39 @@ class GetCurrentTrackAction extends AbstractAction
 		return $this->arguments;
 	}
 
-	private function extendUris( CurrentTrackEntity $currentTrack, StationEntity $station, ?FavoriteEntity $favorite ): void
+	/**
+	 * @param StationEntity[] $stations
+	 */
+	private function extendUris( array $stations ): void
 	{
 		$uriBuilder = $this->getUriBuilder();
-		( new CurrentTrackUriExtender( $uriBuilder, $currentTrack, $station, $favorite ) )
-			->extend();
+		foreach ( $stations as $station )
+		{
+			( new StationUriExtender( $uriBuilder, $station ) )
+				->extend();
+		}
 	}
 
 	/**
 	 * @throws PersistenceException
 	 */
-	private function readStation( StationEntity $requestedStation ): ?StationEntity
-	{
-		$databaseConnector = $this->getDatabaseConnector();
-
-		return ( new StationsRepository( $databaseConnector ) )
-			->readStationById( $requestedStation );
-	}
-
-	private function readCurrentTrack( StationEntity $station ): CurrentTrackEntity
-	{
-		$currentTrack            = new CurrentTrackEntity();
-		$currentTrack->name      = ( new CurrentTrackReader() )
-			->read( $station->tracklistUri, $station->currentTrackXPath );
-		$currentTrack->stationId = $station->id;
-
-		return $currentTrack;
-	}
-
-	/**
-	 * @throws PersistenceException
-	 */
-	private function readFavoriteByTrackName( FavoriteEntity $requestedFavorite ): ?FavoriteEntity
+	private function readFavorite( FavoriteEntity $favorite ): ?FavoriteEntity
 	{
 		$databaseConnector = $this->getDatabaseConnector();
 
 		return ( new FavoritesRepository( $databaseConnector ) )
-			->readFavoriteByTrackName( $requestedFavorite );
+			->readFavoriteById( $favorite );
+	}
+
+	/**
+	 * @return StationEntity[]
+	 * @throws PersistenceException
+	 */
+	private function readFavoriteStations( FavoriteEntity $favorite ): array
+	{
+		$databaseConnector = $this->getDatabaseConnector();
+
+		return ( new StationsRepository( $databaseConnector ) )
+			->readStationsByFavoriteId( $favorite );
 	}
 }
