+++ src/Actions/Api/Get/GetFavoriteUsersAction.php
@@ -1,5 +1,5 @@
 <?php declare( strict_types = 1 );
-namespace CodeKandis\TradioApi\Actions\Api\Read;
+namespace CodeKandis\TradioApi\Actions\Api\Get;
 
 use CodeKandis\Tiphy\Actions\AbstractAction;
 use CodeKandis\Tiphy\Exceptions\ErrorInformation;
@@ -9,17 +9,17 @@ use CodeKandis\Tiphy\Persistence\MariaDb\Connector;
 use CodeKandis\Tiphy\Persistence\MariaDb\ConnectorInterface;
 use CodeKandis\Tiphy\Persistence\PersistenceException;
 use CodeKandis\TradioApi\Configurations\ConfigurationRegistry;
-use CodeKandis\TradioApi\Entities\StationEntity;
+use CodeKandis\TradioApi\Entities\FavoriteEntity;
 use CodeKandis\TradioApi\Entities\UriExtenders\UserUriExtender;
 use CodeKandis\TradioApi\Entities\UserEntity;
-use CodeKandis\TradioApi\Errors\StationsErrorCodes;
-use CodeKandis\TradioApi\Errors\StationsErrorMessages;
+use CodeKandis\TradioApi\Errors\FavoritesErrorCodes;
+use CodeKandis\TradioApi\Errors\FavoritesErrorMessages;
 use CodeKandis\TradioApi\Http\UriBuilders\ApiUriBuilder;
-use CodeKandis\TradioApi\Persistence\MariaDb\Repositories\StationsRepository;
+use CodeKandis\TradioApi\Persistence\MariaDb\Repositories\FavoritesRepository;
 use CodeKandis\TradioApi\Persistence\MariaDb\Repositories\UsersRepository;
 use ReflectionException;
 
-class GetStationUsersAction extends AbstractAction
+class GetFavoriteUsersAction extends AbstractAction
 {
 	/** @var ConnectorInterface */
 	private $databaseConnector;
@@ -57,20 +57,20 @@ class GetStationUsersAction extends AbstractAction
 	{
 		$inputData = $this->getInputData();
 
-		$requestedStation     = new StationEntity();
-		$requestedStation->id = $inputData[ 'id' ];
-		$station              = $this->readStation( $requestedStation );
+		$requestedFavorite     = new FavoriteEntity();
+		$requestedFavorite->id = $inputData[ 'id' ];
+		$favorite              = $this->readFavorite( $requestedFavorite );
 
-		if ( null === $station )
+		if ( null === $favorite )
 		{
-			$errorInformation = new ErrorInformation( StationsErrorCodes::STATION_UNKNOWN, StationsErrorMessages::STATION_UNKNOWN, $inputData );
+			$errorInformation = new ErrorInformation( FavoritesErrorCodes::FAVORITE_UNKNOWN, FavoritesErrorMessages::FAVORITE_UNKNOWN, $inputData );
 			( new JsonResponder( StatusCodes::NOT_FOUND, null, $errorInformation ) )
 				->respond();
 
 			return;
 		}
 
-		$users = $this->readStationUsers( $station );
+		$users = $this->readFavoriteUsers( $favorite );
 		$this->extendUris( $users );
 
 		$responderData = [
@@ -104,23 +104,23 @@ class GetStationUsersAction extends AbstractAction
 	/**
 	 * @throws PersistenceException
 	 */
-	private function readStation( StationEntity $station ): ?StationEntity
+	private function readFavorite( FavoriteEntity $favorite ): ?FavoriteEntity
 	{
 		$databaseConnector = $this->getDatabaseConnector();
 
-		return ( new StationsRepository( $databaseConnector ) )
-			->readStationById( $station );
+		return ( new FavoritesRepository( $databaseConnector ) )
+			->readFavoriteById( $favorite );
 	}
 
 	/**
 	 * @return UserEntity[]
 	 * @throws PersistenceException
 	 */
-	private function readStationUsers( StationEntity $station ): array
+	private function readFavoriteUsers( FavoriteEntity $favorite ): array
 	{
 		$databaseConnector = $this->getDatabaseConnector();
 
 		return ( new UsersRepository( $databaseConnector ) )
-			->readUsersByStationId( $station );
+			->readUsersByFavoriteId( $favorite );
 	}
 }
