+++ src/Actions/Api/Get/GetUserAction.php
@@ -1,5 +1,5 @@
 <?php declare( strict_types = 1 );
-namespace CodeKandis\TradioApi\Actions\Api\Read;
+namespace CodeKandis\TradioApi\Actions\Api\Get;
 
 use CodeKandis\Tiphy\Actions\AbstractAction;
 use CodeKandis\Tiphy\Exceptions\ErrorInformation;
@@ -9,17 +9,15 @@ use CodeKandis\Tiphy\Persistence\MariaDb\Connector;
 use CodeKandis\Tiphy\Persistence\MariaDb\ConnectorInterface;
 use CodeKandis\Tiphy\Persistence\PersistenceException;
 use CodeKandis\TradioApi\Configurations\ConfigurationRegistry;
-use CodeKandis\TradioApi\Entities\StationEntity;
-use CodeKandis\TradioApi\Entities\UriExtenders\StationUriExtender;
+use CodeKandis\TradioApi\Entities\UriExtenders\UserUriExtender;
 use CodeKandis\TradioApi\Entities\UserEntity;
 use CodeKandis\TradioApi\Errors\UsersErrorCodes;
 use CodeKandis\TradioApi\Errors\UsersErrorMessages;
 use CodeKandis\TradioApi\Http\UriBuilders\ApiUriBuilder;
-use CodeKandis\TradioApi\Persistence\MariaDb\Repositories\StationsRepository;
 use CodeKandis\TradioApi\Persistence\MariaDb\Repositories\UsersRepository;
 use ReflectionException;
 
-class GetUserStationsAction extends AbstractAction
+class GetUserAction extends AbstractAction
 {
 	/** @var ConnectorInterface */
 	private $databaseConnector;
@@ -70,11 +68,10 @@ class GetUserStationsAction extends AbstractAction
 			return;
 		}
 
-		$stations = $this->readUsersStations( $user );
-		$this->extendUris( $stations );
+		$this->extendUris( $user );
 
 		$responderData = [
-			'stations' => $stations,
+			'user' => $user
 		];
 		( new JsonResponder( StatusCodes::OK, $responderData ) )
 			->respond();
@@ -88,17 +85,11 @@ class GetUserStationsAction extends AbstractAction
 		return $this->arguments;
 	}
 
-	/**
-	 * @param StationEntity[] $stations
-	 */
-	private function extendUris( array $stations ): void
+	private function extendUris( UserEntity $user ): void
 	{
 		$uriBuilder = $this->getUriBuilder();
-		foreach ( $stations as $station )
-		{
-			( new StationUriExtender( $uriBuilder, $station ) )
-				->extend();
-		}
+		( new UserUriExtender( $uriBuilder, $user ) )
+			->extend();
 	}
 
 	/**
@@ -110,17 +101,5 @@ class GetUserStationsAction extends AbstractAction
 
 		return ( new UsersRepository( $databaseConnector ) )
 			->readUserById( $requestedUser );
-	}
-
-	/**
-	 * @return StationEntity[]
-	 * @throws PersistenceException
-	 */
-	private function readUsersStations( UserEntity $user ): array
-	{
-		$databaseConnector = $this->getDatabaseConnector();
-
-		return ( new StationsRepository( $databaseConnector ) )
-			->readStationsByUserId( $user );
 	}
 }
